Initialize pom.xml with basic setup and corresponding examples
The following dependencies are included so far:

- Java EE 7.0
- JSF 2.2
- Glassfish 4.1.1
- Junit 4.12

The following maven plugins have been included:

- maven-compiler-plugin (allows specification of java version)
- embedded glassfish plugin (easy to use glassfish plugin)
- maven war plugin (for packaging exmartikulator as a .war file)

For packaging .war seems to be a requirement because certified Java EE
applications servers (like glassfish) are not able to deploy .jar files.
The alternative is the EJB packaging format but it seems to be even more
bloated than the WAR format.

We use the maven embedded glassfish plugin instead of the 'standard'
maven glassfish plugin because the former makes it easier to get
glassfish up and running without shooting yourself in the foot. Since
the formal requirements for the SWP2 project haven't been published so
far we might be forced to change this in the future.

Besides the pom.xml a boilerplate class, a sample junit test and a
.gitignore file are included in this commit. In addition to that I also
wrote a sample web.xml and index.xhtml file.

To get this sample up and running the following steps have to be
performed: 1) A .war package for Exmartikulator needs to be created and
2) The developer needs to start glassfish.

Both steps can be performed with the following command:

	$ mvn package embedded-glassfish:run

Afterwards go to: http://0.0.0.0:8080/exmartikulator/index.html