Create a CONTRIBUTING file to clarify contribution guidelines
The file explains how to make new changes and how to create new merge
requests. It also describes the process of merging proposed changes and
defines style guidelines for XML and Java files. More style guides might
be added in the future. In addition to that a small section on good
commit messages is included as well.