Add configuration file for GitLab CI
This configuration files differentiates between 3 stages (build, test
and verify). The build stage does what you expect it to do: it builds
the entire thing. The test stage invokes all surefire tests by executing
the cobertura:check goal thereby also ensuring that our test coverage is
above a configured minimum. In the future it will take a lot of time to
run all tests and perform test coverage analysis which is the reason why
these first two stages are only invoked on the master branch.

The last stage (verify) is invoked on all branches. It performs static
code analysis. To be more exact it runs findBugs:check and
checkstyle:check it is executed on all branches.

Currently mvn is only installed on the Linux computers, however, we
could manually install maven on the Mac OS X and sunos computers. We
might do so in the future...