Initialize TME parser
The format itself is documented in the various methods of the Parser
class (as an EBNF). The parsing code is pretty much finished the only
thing that's still missing is a function for parsing the hash syntax,
however, the hash syntax is not very clear from the given examples. I
need to ask Marcel if he could clarify the hash syntax before working on
a parser function for it.

The javadoc and the tests could be improved further, but for now I
believe that they are good enough.