Backend logic for password reset functionality
This is the basic backend logic for a password reset function. It
doesn't send emails yet because the mail code is not yet merged (see
!28), however everything else is basically working. None the less I
cannot guarantee that it is 100% bullet proof because we don't have any
tests, so if it breaks in the future don't blame me.

Each user can only have a single password reset token at a time. The
token expires after a certain amount of time and if it has expired a new
token can be created. In the future we might allow users to mark tokens
as expired manually (in case someone else has triggered a password
reset) and we could consider allowing more than one token at a time, but
for now it's good enough as it is.

Fixes #26