Implement removal of tutorials using the modal removal dialog
There is one problem with this approach: Even though the `update`
attribute of the `commandLink` Node is set correctly you are required
two refresh the page in order to get an updated Tutorial List.

No idea what's causing this, maybe it's cached somewhere? Anyways, the
dirty solution would be to force refresh the package after closing the
modal dialog.