Refactor Tutorial controller by using a DualModelList
This requires a converter now, further more to share the course context
between the converter and the controller I added the course to the
session map. We might want to always have the current course in the
session map (this has to be implemented in the filter).

Furthermore the converter has not been moved to a separate package
because it's a @ManagedBean and not a @FacesConverter.